# server/swagger.json
{"swagger":"2.0","info":{"title":"API Docs","version":"1.0.0"},"basePath":"/","paths":{"/events":{"post":{"tags":["event"],"description":"","responses":{"200":{"description":"Operation finished successfully."},"400":{"description":"Invalid parameter(s). (More details on response object)"},"500":{"description":"Server error. (More details on response object)"}},"parameters":[{"name":"x-access-token","in":"header","description":"JWT created on user creation or authentication.","type":"string","required":true},{"name":"event","in":"body","description":"JSON representation of the event to be created.","schema":{"type":"object"},"required":false}],"operationId":"saveEvent","addedParameters":{"header":{"x-access-token":true},"body":{"event":true}}}},"/events/by-user":{"get":{"tags":["event"],"description":"","responses":{"200":{"description":"Operation finished successfully."},"400":{"description":"Invalid parameter(s). (More details on response object)"},"500":{"description":"Server error. (More details on response object)"}},"parameters":[{"name":"x-access-token","in":"header","description":"JWT created on user creation or authentication.","type":"string","required":true}],"operationId":"getEvent","addedParameters":{"header":{"x-access-token":true}}}},"/events/{id}":{"get":{"tags":["event"],"description":"","responses":{"200":{"description":"Operation finished successfully."},"400":{"description":"Invalid parameter(s). (More details on response object)"},"500":{"description":"Server error. (More details on response object)"}},"parameters":[{"name":"x-access-token","in":"header","description":"JWT created on user creation or authentication.","type":"string","required":true},{"name":"id","in":"path","type":"string","required":true}],"operationId":"getEvent","addedParameters":{"header":{"x-access-token":true},"path":{"id":true}}}},"/events/deactivate/{id}":{"post":{"tags":["event"],"description":"","responses":{"200":{"description":"Operation finished successfully."},"400":{"description":"Invalid parameter(s). (More details on response object)"},"500":{"description":"Server error. (More details on response object)"}},"parameters":[{"name":"x-access-token","in":"header","description":"JWT created on user creation or authentication.","type":"string","required":true},{"name":"id","in":"path","type":"string","required":true}],"operationId":"deactivateEvent","addedParameters":{"header":{"x-access-token":true},"path":{"id":true}}}},"/events/activate/{id}":{"post":{"tags":["event"],"description":"","responses":{"200":{"description":"Operation finished successfully."},"400":{"description":"Invalid parameter(s). (More details on response object)"},"500":{"description":"Server error. (More details on response object)"}},"parameters":[{"name":"x-access-token","in":"header","description":"JWT created on user creation or authentication.","type":"string","required":true},{"name":"id","in":"path","type":"string","required":true}],"operationId":"activateEvent","addedParameters":{"header":{"x-access-token":true},"path":{"id":true}}}},"/users":{"post":{"tags":["user"],"description":"","responses":{"200":{"description":"Operation finished successfully."},"400":{"description":"Invalid parameter(s). (More details on response object)"},"500":{"description":"Server error. (More details on response object)"}},"parameters":[{"name":"user","in":"body","description":"JSON representation of the user to be created.","schema":{"type":"object"},"required":false}],"operationId":"createUser","addedParameters":{"body":{"user":true}}},"put":{"tags":["user"],"description":"","responses":{"200":{"description":"Operation finished successfully."},"400":{"description":"Invalid parameter(s). (More details on response object)"},"500":{"description":"Server error. (More details on response object)"}},"parameters":[{"name":"x-access-token","in":"header","description":"JWT created on user creation or authentication.","type":"string","required":true},{"name":"user","in":"body","description":"JSON representation of the user to be created.","schema":{"type":"object"},"required":false}],"operationId":"updateUser","addedParameters":{"header":{"x-access-token":true},"body":{"user":true}}}},"/users/profile":{"get":{"tags":["user"],"description":"","responses":{"200":{"description":"Operation finished successfully."},"400":{"description":"Invalid parameter(s). (More details on response object)"},"500":{"description":"Server error. (More details on response object)"}},"parameters":[{"name":"x-access-token","in":"header","description":"JWT created on user creation or authentication.","type":"string","required":true}],"operationId":"getUserData","addedParameters":{"header":{"x-access-token":true}}}},"/users/username/exists/{username}":{"get":{"tags":["user"],"description":"","responses":{"200":{"description":"Operation finished successfully."},"400":{"description":"Invalid parameter(s). (More details on response object)"},"500":{"description":"Server error. (More details on response object)"}},"parameters":[{"name":"username","in":"path","type":"string","required":true}],"operationId":"isUniqueUsername","addedParameters":{"path":{"username":true}}}},"/users/authenticate":{"post":{"tags":["user"],"description":"","responses":{"200":{"description":"Operation finished successfully."},"400":{"description":"Invalid parameter(s). (More details on response object)"},"500":{"description":"Server error. (More details on response object)"}},"parameters":[{"name":"user","in":"body","description":"User credentials object.","schema":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"required":false}],"operationId":"authenticateUser","addedParameters":{"body":{"user":true}}}},"/users/password/recover":{"post":{"tags":["user"],"description":"","responses":{"200":{"description":"Operation finished successfully."},"400":{"description":"Invalid parameter(s). (More details on response object)"},"500":{"description":"Server error. (More details on response object)"}},"parameters":[{"name":"user","in":"body","description":"User e-mail.","schema":{"type":"object","properties":{"email":{"type":"string"}},"required":["email"]},"required":false}],"operationId":"recoverPassword","addedParameters":{"body":{"user":true}}}},"/users/password/update":{"post":{"tags":["user"],"description":"","responses":{"200":{"description":"Operation finished successfully."},"400":{"description":"Invalid parameter(s). (More details on response object)"},"500":{"description":"Server error. (More details on response object)"}},"parameters":[{"name":"x-access-token","in":"header","description":"JWT created on user creation or authentication.","type":"string","required":true},{"name":"user","in":"body","description":"User credentials object.","schema":{"type":"object","properties":{"password":{"type":"string"},"newPassword":{"type":"string"}},"required":["password","newPassword"]},"required":false}],"operationId":"updatePassword","addedParameters":{"header":{"x-access-token":true},"body":{"user":true}}}}}}